Remove the tailing C2R for GpuDataWritingCommandExec.
Spark will append a ColumnarToRow (C2R) to a GpuDataWritingCommandExec, but this
C2R is useless because DataWritingCommandExec returns empty data.

Besides, Spark will trigger a new job for the operators after a DataWritingCommandExec,
and makes the DataWritingCommandExec be the data source node for this new job. Since
no data is generated, this new job is entirely a noop.

Removing the tailing C2R can eliminate the useless job. And then GpuDataWritingCommandExec
becomes the root node, and should support row-based executions to interact with Spark.
